pyproject.toml

[project]
name = "pubmed-article-search"
version = "0.1.0"
description = "A tool to search and retrieve PubMed articles, filter by author and topic, and export results to CSV"
authors = [
    {name = "Saikat Sinha", email = "saikatsinha21@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
pandas = "^1.3"
biopython = "^1.79"
streamlit = "^1.15"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"